GH-45563: [C++][Compute] Split up hash_aggregate.cc
